requirements.txt
python-dotenv==1.2.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.30
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
openai==2.16.0
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.1.0
pandas==3.0.2
numpy==2.4.4
streamlit==1.57.0
plotly==6.7.0
apscheduler==3.11.2
pytest==8.2.0
pytest-asyncio==0.23.6
